refactor(cordis): @dsh-hanako 子插件改名——api-bridge→bridge / bridge→bus / web-app→app
- bridge: 数据面桥（免鉴权 connection 等价服务 + /api HTTP 载体，原 api-bridge）
- bus: 消息总线（dshana.bus WS 服务端 + RPC 翻译器 + 事件流，原 bridge；与宿主侧
  src/lib/bus.js 客户端对称命名）
- app: WebUI 应用载体（serve 官方 dist 到根路径，原 web-app）
- 同步: cordis.patch.yml 挂载 id/name、package.json、client.js 模块 id、代码注释
