--- src-cordis/plugins/@dsh-hanako/api-bridge/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@dsh-hanako/api-bridge",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "main": "index.js",
-  "exports": {
-    ".": "./index.js",
-    "./client": "./client.js",
-    "./package.json": "./package.json"
-  },
-  "dsh": {
-    "client": {
-      "inject": [],
-      "platform": "web",
-      "immediately": true
-    }
-  },
-  "cordis": {
-    "name": "@dsh-hanako/api-bridge"
-  }
-}
