+++ src-cordis/plugins/@dsh-hanako/bus/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@dsh-hanako/bus",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "main": "index.js",
+  "cordis": {
+    "name": "@dsh-hanako/bus"
+  }
+}
